Changed check functions and fixed the queue error
Instead of returning +1,-1,0 the check_gameover function is now designed to return the difference, which is the maximum absolute difference across all rows, columns, and diagonals...essentially the best chance to win and who has that chance(+,-). 

There were 2 more errors: 
1)The parent's is_max was directly assigned to the children, which is obviously wrong, we should invert it. Fixed the issue.
2) Once a gameover state entered the queue, instead of pushing it over and continuing to expand the remaining states, the algorithm halted there. 
This was because we popped from the queue only if it wasn't a end state, which is obviously wrong. We should pop it anyway regardless of whether its a leaf or not